fib3

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,9 +6,9 @@
  */
 int main(void)
 {
-	double un = 1;
-	double un_p1 = un;
-	double un_m1 = 0;
+	long double un = 1;
+	long double un_p1 = un;
+	long double un_m1 = 0;
 	unsigned long n = 1;
 
 	while (n <= 98)
@@ -16,7 +16,7 @@ int main(void)
 		un_p1 = un + un_m1;
 		un_m1 = un;
 		un = un_p1;
-		printf("%.0f", un_p1);
+		printf("%.0Lf", un_p1);
 		if (n < 98)
 			printf(", ");
 		n++;
